requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.2
coverage==7.13.4
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.0
filetype==1.2.0
fonttools==4.61.1
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.6.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-core==1.2.17
langchain-google-genai==4.2.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.14
lazy-loader==0.5
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
onnx==1.20.1
openai==2.26.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
optimum==2.1.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
pdf2image==1.17.0
pillow==12.1.1
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
portalocker==3.2.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.17.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2025.5.10
together==2.3.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
types-tabulate==0.10.0.20260308
types-tqdm==4.67.3.20260303
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
